requirements.txt
annotated-types
anyio
certifi
colorama
distro
h11
httpcore
httpx
idna
jiter
openai
pydantic
pydantic_core
python-dotenv
sniffio
tqdm
typing-inspection
typing_extensions
pillow
rich
requests
python-jose[cryptography]

# fastapi
# uvicorn
# python-multipart
# sqlalchemy
# passlib[bcrypt]
# aspose-words
# openai
# python-dotenv
# pydantic[email]
# translitua